fix(frontend): move download report from topbar to sidebar
