+++ webapp/src/language/en.json
@@ -34,6 +34,8 @@
       "homepage": "Home page",
       "tools": "Tools",
       "yourDonations": "Your Donations",
+      "downloadReport": "Download report",
+      "editPage": "Edit page",
       "information": "Information",
       "aboutLifebank": "About Lifebank",
       "useTerms": "Terms of Use",
