+++ webapp/src/language/es.json
@@ -34,6 +34,8 @@
       "homepage": "Página de inicio",
       "tools": "Herramientas",
       "yourDonations": "Tus donaciones",
+      "downloadReport": "Descargar reporte",
+      "editPage": "Edit página",
       "information": "Información",
       "aboutLifebank": "Acerca de Lifebank",
       "useTerms": "Términos de uso",
